+++ vllm_gguf_plugin/triton/fused_moe/iq_quant/iq3_s.py
@@ -1,0 +1,163 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+import torch
+import triton
+import triton.language as tl
+
+from ...gemm.iq_quant.iq_tables import get_iq_table_tensors
+from ...gemm.utils import (
+    GGML_TYPE_IQ1_M,
+    GGML_TYPE_IQ1_S,
+    GGML_TYPE_IQ2_S,
+    GGML_TYPE_IQ2_XXS,
+    GGML_TYPE_IQ2_XS,
+    GGML_TYPE_IQ3_S,
+    GGML_TYPE_IQ3_XXS,
+    GGML_TYPE_IQ4_NL,
+    GGML_TYPE_IQ4_XS,
+    load_f16_from_u8,
+    load_u16_from_u8,
+    load_u32_from_u8,
+)
+from ..utils import (
+    load_moe_token_info,
+    load_moe_x_chunk,
+    load_moe_x_tile,
+    run_triton_fused_moe_kernel,
+)
+
+
+@triton.jit
+def iq3_s_moe_kernel(
+    x_ptr,
+    w_u8_ptr,
+    y_ptr,
+    sorted_token_ids_ptr,
+    expert_ids_ptr,
+    num_tokens_post_padded_ptr,
+    num_valid_tokens,
+    top_k,
+    n,
+    num_k_blocks,
+    stride_xm,
+    stride_xk,
+    stride_we,
+    stride_wn,
+    stride_wk,
+    stride_ym,
+    stride_yn,
+    grid_ptr,
+    BLOCK_M: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+    BLOCK_K_BLOCKS: tl.constexpr,
+):
+    pid_m = tl.program_id(axis=0)
+    pid_n = tl.program_id(axis=1)
+    offs_output, offs_token, token_mask = load_moe_token_info(
+        sorted_token_ids_ptr, pid_m, top_k, num_valid_tokens, BLOCK_M=BLOCK_M
+    )
+    expert = tl.load(expert_ids_ptr + pid_m)
+    if expert < 0:
+        return
+
+    offs_n = pid_n * BLOCK_N + tl.arange(0, BLOCK_N)
+    offs_4 = tl.arange(0, 4)
+    sign_mask_lo = (1 << offs_4).to(tl.uint8)
+    sign_mask_hi = (1 << (offs_4 + 4)).to(tl.uint8)
+    n_mask = offs_n < n
+
+    acc = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
+    w_row_ptrs = w_u8_ptr + expert * stride_we + offs_n * stride_wn
+
+    for kb in range(0, num_k_blocks):
+        block_ptrs = w_row_ptrs + kb * 110
+
+        for ib in range(8):
+            qh = tl.load(block_ptrs + 66 + ib, mask=n_mask, other=0).to(tl.int32)
+            scale_byte = tl.load(block_ptrs + 106 + (ib // 2), mask=n_mask, other=0)
+            for il in range(4):
+                signs = tl.load(block_ptrs + 74 + 4 * ib + il, mask=n_mask, other=0)
+                qs_base = block_ptrs + 2 + 8 * ib
+                idx1 = tl.load(qs_base + 2 * il + 0, mask=n_mask, other=0).to(tl.int32) | (
+                    (qh << (8 - 2 * il)) & 0x100
+                )
+                idx2 = tl.load(qs_base + 2 * il + 1, mask=n_mask, other=0).to(tl.int32) | (
+                    (qh << (7 - 2 * il)) & 0x100
+                )
+                x1 = load_moe_x_chunk(
+                    x_ptr,
+                    stride_xm,
+                    stride_xk,
+                    offs_token,
+                    token_mask,
+                    kb * 256 + 32 * ib + 8 * il,
+                    CHUNK=4,
+                )
+                x2 = load_moe_x_chunk(
+                    x_ptr,
+                    stride_xm,
+                    stride_xk,
+                    offs_token,
+                    token_mask,
+                    kb * 256 + 32 * ib + 8 * il + 4,
+                    CHUNK=4,
+                )
+                x_dtype = x1.dtype
+                d = load_f16_from_u8(block_ptrs + 0, n_mask).to(x_dtype)
+                dscale = (
+                    d * (1 + 2 * ((scale_byte >> (4 * (ib % 2))) & 0x0F).to(x_dtype))
+                ).to(x_dtype)
+                grid1 = tl.load(
+                    grid_ptr + idx1[:, None] * 4 + offs_4[None, :],
+                    mask=n_mask[:, None],
+                    other=0,
+                ).to(x_dtype)
+                grid2 = tl.load(
+                    grid_ptr + idx2[:, None] * 4 + offs_4[None, :],
+                    mask=n_mask[:, None],
+                    other=0,
+                ).to(x_dtype)
+                q1 = (
+                    grid1
+                    * tl.where((signs[:, None] & sign_mask_lo[None, :]) != 0, -1, 1).to(x_dtype)
+                    * dscale[:, None]
+                ).to(x_dtype)
+                q2 = (
+                    grid2
+                    * tl.where((signs[:, None] & sign_mask_hi[None, :]) != 0, -1, 1).to(x_dtype)
+                    * dscale[:, None]
+                ).to(x_dtype)
+                acc += tl.sum(x1[:, None, :] * q1[None, :, :], axis=2)
+                acc += tl.sum(x2[:, None, :] * q2[None, :, :], axis=2)
+
+    y_ptrs = y_ptr + offs_output[:, None] * stride_ym + offs_n[None, :] * stride_yn
+    y_mask = token_mask[:, None] & n_mask[None, :]
+    tl.store(y_ptrs, acc, mask=y_mask)
+
+
+def ggml_moe_iq3_s_triton(
+    X: torch.Tensor,
+    W: torch.Tensor,
+    sorted_token_ids: torch.Tensor,
+    expert_ids: torch.Tensor,
+    num_tokens_post_padded: torch.Tensor,
+    row: int,
+    top_k: int,
+    tokens: int,
+) -> torch.Tensor:
+    tables = get_iq_table_tensors(W.device)
+    return run_triton_fused_moe_kernel(
+        iq3_s_moe_kernel,
+        W,
+        X,
+        sorted_token_ids,
+        expert_ids,
+        num_tokens_post_padded,
+        row,
+        top_k,
+        tokens,
+        GGML_TYPE_IQ3_S,
+        extra_args=(tables["iq3s_grid"],),
+    )
