+++ vllm_gguf_plugin/triton/fused_moe/k_quant/q3_k.py
@@ -1,0 +1,135 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+import torch
+import triton
+import triton.language as tl
+
+from ...gemm.utils import (
+    GGML_TYPE_Q2_K,
+    GGML_TYPE_Q3_K,
+    GGML_TYPE_Q4_K,
+    GGML_TYPE_Q5_K,
+    GGML_TYPE_Q6_K,
+    load_f16_from_u8,
+)
+from ..utils import (
+    load_moe_token_info,
+    load_moe_x_chunk,
+    run_triton_fused_moe_kernel,
+)
+
+
+@triton.jit
+def q3_k_moe_kernel(
+    x_ptr,
+    w_u8_ptr,
+    y_ptr,
+    sorted_token_ids_ptr,
+    expert_ids_ptr,
+    num_tokens_post_padded_ptr,
+    num_valid_tokens,
+    top_k,
+    n,
+    num_k_blocks,
+    stride_xm,
+    stride_xk,
+    stride_we,
+    stride_wn,
+    stride_wk,
+    stride_ym,
+    stride_yn,
+    BLOCK_M: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+    BLOCK_K_BLOCKS: tl.constexpr,
+):
+    pid_m = tl.program_id(axis=0)
+    pid_n = tl.program_id(axis=1)
+    offs_output, offs_token, token_mask = load_moe_token_info(
+        sorted_token_ids_ptr, pid_m, top_k, num_valid_tokens, BLOCK_M=BLOCK_M
+    )
+    expert = tl.load(expert_ids_ptr + pid_m)
+    if expert < 0:
+        return
+
+    offs_n = pid_n * BLOCK_N + tl.arange(0, BLOCK_N)
+    offs_l = tl.arange(0, 16)
+    n_mask = offs_n < n
+
+    acc = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
+    w_row_ptrs = w_u8_ptr + expert * stride_we + offs_n * stride_wn
+
+    for kb in range(0, num_k_blocks):
+        block_ptrs = w_row_ptrs + kb * 110
+        scales_ptr = block_ptrs + 96
+
+        for chunk in range(16):
+            group = chunk // 8
+            rem = chunk % 8
+            j = rem // 2
+            half = rem % 2
+            hm_mask = 1 << (4 * group + j)
+            is_idx = 8 * group + 2 * j + half
+            if is_idx < 4:
+                lo = tl.load(scales_ptr + is_idx, mask=n_mask, other=0)
+                hi = tl.load(scales_ptr + is_idx + 8, mask=n_mask, other=0)
+                scale = (lo & 0x0F) | (((hi >> 0) & 0x03) << 4)
+            elif is_idx < 8:
+                lo = tl.load(scales_ptr + is_idx, mask=n_mask, other=0)
+                hi = tl.load(scales_ptr + is_idx + 4, mask=n_mask, other=0)
+                scale = (lo & 0x0F) | (((hi >> 2) & 0x03) << 4)
+            elif is_idx < 12:
+                lo = tl.load(scales_ptr + is_idx - 8, mask=n_mask, other=0)
+                hi = tl.load(scales_ptr + is_idx, mask=n_mask, other=0)
+                scale = ((lo >> 4) & 0x0F) | (((hi >> 4) & 0x03) << 4)
+            else:
+                lo = tl.load(scales_ptr + is_idx - 8, mask=n_mask, other=0)
+                hi = tl.load(scales_ptr + is_idx - 4, mask=n_mask, other=0)
+                scale = ((lo >> 4) & 0x0F) | (((hi >> 6) & 0x03) << 4)
+
+            x_tile = load_moe_x_chunk(
+                x_ptr,
+                stride_xm,
+                stride_xk,
+                offs_token,
+                token_mask,
+                kb * 256 + 128 * group + 32 * j + 16 * half,
+                CHUNK=16,
+            )
+            x_dtype = x_tile.dtype
+            d = load_f16_from_u8(block_ptrs + 108, n_mask).to(x_dtype)
+            ql = tl.load(
+                block_ptrs[:, None] + 32 + 32 * group + 16 * half + offs_l[None, :],
+                mask=n_mask[:, None],
+                other=0,
+            )
+            hm = tl.load(
+                block_ptrs[:, None] + 16 * half + offs_l[None, :],
+                mask=n_mask[:, None],
+                other=0,
+            )
+            q = ((ql >> (2 * j)) & 0x03).to(tl.int16)
+            q = (q - tl.where((hm & hm_mask) != 0, 0, 4).to(tl.int16)).to(x_dtype)
+            dq = ((scale.to(tl.int16) - 32).to(x_dtype) * d)[:, None]
+            q_tile = q * dq
+            acc = tl.dot(x_tile, tl.trans(q_tile), acc=acc)
+
+    y_ptrs = y_ptr + offs_output[:, None] * stride_ym + offs_n[None, :] * stride_yn
+    y_mask = token_mask[:, None] & n_mask[None, :]
+    tl.store(y_ptrs, acc, mask=y_mask)
+
+
+def ggml_moe_q3_k_triton(
+    X: torch.Tensor,
+    W: torch.Tensor,
+    sorted_token_ids: torch.Tensor,
+    expert_ids: torch.Tensor,
+    num_tokens_post_padded: torch.Tensor,
+    row: int,
+    top_k: int,
+    tokens: int,
+) -> torch.Tensor:
+    return run_triton_fused_moe_kernel(
+        q3_k_moe_kernel, W, X, sorted_token_ids, expert_ids, num_tokens_post_padded, row, top_k, tokens, GGML_TYPE_Q3_K
+    )
