+++ vllm_gguf_plugin/triton/fused_moe/standard_quant/__init__.py
@@ -1,0 +1,17 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from .q4_0 import ggml_moe_q4_0_triton
+from .q4_1 import ggml_moe_q4_1_triton
+from .q5_0 import ggml_moe_q5_0_triton
+from .q5_1 import ggml_moe_q5_1_triton
+from .q8_0 import ggml_moe_q8_0_triton
+from .q8_1 import ggml_moe_q8_1_triton
+
+__all__ = [
+    "ggml_moe_q4_0_triton",
+    "ggml_moe_q4_1_triton",
+    "ggml_moe_q5_0_triton",
+    "ggml_moe_q5_1_triton",
+    "ggml_moe_q8_0_triton",
+    "ggml_moe_q8_1_triton",
+]
